requirements.txt
asgiref==3.6.0
dj-database-url==1.3.0
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.5.0
